+++ src/com/intellij/idea/plugin/hybris/project/configurators/VersionControlSystemConfigurator.kt
@@ -18,16 +18,38 @@
  */
 package com.intellij.idea.plugin.hybris.project.configurators
 
+import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.vcs.ProjectLevelVcsManager
+import com.intellij.openapi.vcs.VcsDirectoryMapping
+import com.intellij.openapi.vcs.roots.VcsRootDetector
+import com.intellij.openapi.vfs.VfsUtil
 
-interface VersionControlSystemConfigurator {
+@Service
+class VersionControlSystemConfigurator {
 
     fun configure(
         indicator: ProgressIndicator,
         hybrisProjectDescriptor: HybrisProjectDescriptor,
         project: Project
-    )
+    ) {
+        indicator.text = HybrisI18NBundleUtils.message("hybris.project.import.vcs")
+
+        val vcsManager = ProjectLevelVcsManager.getInstance(project)
+        val rootDetector = project.getService(VcsRootDetector::class.java)
+        val detectedRoots = HashSet(rootDetector.detect())
+
+        val roots = hybrisProjectDescriptor.detectedVcs
+            .mapNotNull { VfsUtil.findFileByIoFile(it, true) }
+            .flatMap { rootDetector.detect(it) }
+        detectedRoots.addAll(roots)
+
+        vcsManager.directoryMappings = detectedRoots
+            .filter { it.vcs != null }
+            .map { VcsDirectoryMapping(it.path.path, it.vcs!!.name) }
+    }
 
 }