+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/CCv2Listeners.kt
@@ -25,20 +25,20 @@ import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Deployment
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Environment
 
 sealed interface CCv2Listener<T : CCv2DTO> {
-    fun fetchingStarted() = Unit
-    fun fetchingCompleted(data: Map<CCv2Subscription, Collection<T>> = emptyMap()) = Unit
+    fun onFetchingStarted(subscriptions: Collection<CCv2Subscription>) = Unit
+    fun onFetchingCompleted(data: Map<CCv2Subscription, Collection<T>> = emptyMap()) = Unit
 }
 
 interface CCv2SettingsListener {
-    fun subscriptionsChanged(subscriptions: List<CCv2Subscription>) = Unit
+    fun onSubscriptionsChanged(subscriptions: List<CCv2Subscription>) = Unit
 }
 
 interface CCv2EnvironmentsListener : CCv2Listener<CCv2Environment>
 interface CCv2DeploymentsListener : CCv2Listener<CCv2Deployment>
 
 interface CCv2BuildsListener : CCv2Listener<CCv2Build> {
-    fun buildStarted() = Unit
-    fun buildRequested(subscription: CCv2Subscription, build: CCv2Build? = null) = Unit
-    fun buildRemovalStarted(subscription: CCv2Subscription, build: CCv2Build) = Unit
-    fun buildRemovalRequested(subscription: CCv2Subscription, build: CCv2Build) = Unit
+    fun onBuildStarted() = Unit
+    fun onBuildRequested(subscription: CCv2Subscription, build: CCv2Build? = null) = Unit
+    fun onBuildRemovalStarted(subscription: CCv2Subscription, build: CCv2Build) = Unit
+    fun onBuildRemovalRequested(subscription: CCv2Subscription, build: CCv2Build) = Unit
 }