+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/actions/AbstractCCv2Action.kt
@@ -23,8 +23,8 @@ import com.intellij.idea.plugin.hybris.toolwindow.HybrisToolWindowFactory
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2Tab
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2View
 import com.intellij.openapi.actionSystem.ActionUpdateThread
-import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindowManager
 import javax.swing.Icon
@@ -34,7 +34,7 @@ abstract class AbstractCCv2Action(
     text: String,
     description: String? = null,
     icon: Icon
-) : AnAction(text, description, icon) {
+) : DumbAwareAction(text, description, icon) {
 
     override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT